feat: Final version run with 100% of data
This is the final version of the project to be submitted. It was run
with 100% of the data to train the model and save the pickle file to be
used by the application.

I have added also the notebooks used to deploy the application as well
as a .gitignore file to not load extra data to teh Github repository.